+++ src/sciform/options/finalized_options.py
@@ -19,9 +19,9 @@ class FinalizedOptions:
     """Rendered options: All options populated and using Enum instead of Literal."""
 
     exp_mode: option_types.ExpModeEnum
-    exp_val: int | type(option_types.AutoExpVal)
+    exp_val: int | option_types.ExpValEnum
     round_mode: option_types.RoundModeEnum
-    ndigits: int | type(option_types.AutoDigits)
+    ndigits: int | option_types.NDigitsEnum
     upper_separator: option_types.UpperSeparatorEnums
     decimal_separator: option_types.DecimalSeparatorEnums
     lower_separator: option_types.LowerSeparatorEnums
@@ -36,7 +36,6 @@ class FinalizedOptions:
     superscript: bool
     nan_inf_exp: bool
     paren_uncertainty: bool
-    pdg_sig_figs: bool
     left_pad_matching: bool
     paren_uncertainty_trim: bool
     pm_whitespace: bool