+++ src/sciform/options/validation.py
@@ -28,6 +28,7 @@ def validate_options(
 
 
 allowed_round_modes = get_args(option_types.RoundMode)
+allowed_ndigits = get_args(option_types.NDigits)
 
 
 def validate_rounding(
@@ -38,8 +39,10 @@ def validate_rounding(
     r"""Validate ndigits if round_mode == "sig_fig"."""
     if none_allowed and options.ndigits is None:
         pass
-    elif not isinstance(options.ndigits, int) and options.ndigits != "all":
-        msg = f"ndigits must be an int or 'all', not {options.ndigits}."
+    elif (
+        not isinstance(options.ndigits, int) and options.ndigits not in allowed_ndigits
+    ):
+        msg = f"ndigits must be an int or in {allowed_ndigits}, not {options.ndigits}."
         raise TypeError(msg)
 
     if none_allowed and options.round_mode is None: